chore: move modals to own directory
