+++ front/src/components/LexicalEditor/components/modals/TableModal.tsx
@@ -2,14 +2,14 @@ import React from 'react';
 
 import { Input } from '@/components/tailwind-ui';
 
-import { useInsertModalContext } from './InsertOptionsMenu';
+import { useInsertModalContext } from '../InsertOptionsMenu';
 
 export function TableModal() {
   const { state, setState } = useInsertModalContext();
   const { rows, columns } =
     (state as Record<'rows' | 'columns', string> | null) ?? {};
   return (
-    <div className="m-2 min-w-1/4">
+    <div className="min-w-1/4 m-2 min-h-[200px]">
       <Input
         label="Rows"
         name="rows"
